refactor: consolidate HTTP date formatting into cachecontrol.HTTPDate
Replace the duplicated t.UTC().Format(http.TimeFormat) expression in the bindings helper and the rate limiter with a single shared cachecontrol.HTTPDate.
