+++ app/transports/http/bindings/accounts.go
@@ -116,7 +116,7 @@ func (i *Accounts) AccountGet(ctx context.Context, request openapi.AccountGetReq
 		return openapi.AccountGet304Response{
 			Headers: openapi.NotModifiedResponseHeaders{
 				CacheControl: accountGetCacheControl,
-				LastModified: httpDate(etag.Time),
+				LastModified: cachecontrol.HTTPDate(etag.Time),
 				ETag:         etag.String(),
 			},
 		}, nil
@@ -137,7 +137,7 @@ func (i *Accounts) AccountGet(ctx context.Context, request openapi.AccountGetReq
 			Body: serialiseAccount(acc),
 			Headers: openapi.AccountGetOKResponseHeaders{
 				CacheControl: accountGetCacheControl,
-				LastModified: httpDate(etag.Time),
+				LastModified: cachecontrol.HTTPDate(etag.Time),
 				ETag:         etag.String(),
 			},
 		},
@@ -160,7 +160,7 @@ func (i *Accounts) AccountView(ctx context.Context, request openapi.AccountViewR
 			Body: serialiseAccount(acc),
 			Headers: openapi.AccountGetOKResponseHeaders{
 				CacheControl: "private, no-cache",
-				LastModified: httpDate(acc.UpdatedAt),
+				LastModified: cachecontrol.HTTPDate(acc.UpdatedAt),
 			},
 		},
 	}, nil
@@ -208,7 +208,7 @@ func (i *Accounts) AccountManageCreate(ctx context.Context, request openapi.Acco
 			Body: serialiseAccount(acc),
 			Headers: openapi.AccountGetOKResponseHeaders{
 				CacheControl: "private, no-cache",
-				LastModified: httpDate(acc.UpdatedAt),
+				LastModified: cachecontrol.HTTPDate(acc.UpdatedAt),
 			},
 		},
 	}, nil
@@ -250,7 +250,7 @@ func (i *Accounts) AccountManageUpdate(ctx context.Context, request openapi.Acco
 			Body: serialiseAccount(acc),
 			Headers: openapi.AccountGetOKResponseHeaders{
 				CacheControl: "private, no-cache",
-				LastModified: httpDate(acc.UpdatedAt),
+				LastModified: cachecontrol.HTTPDate(acc.UpdatedAt),
 			},
 		},
 	}, nil
