+++ app/transports/http/bindings/categories.go
@@ -85,7 +85,7 @@ func (c Categories) CategoryGet(ctx context.Context, request openapi.CategoryGet
 		return openapi.CategoryGet304Response{
 			Headers: openapi.NotModifiedResponseHeaders{
 				CacheControl: categoryGetCacheControl,
-				LastModified: httpDate(etag.Time),
+				LastModified: cachecontrol.HTTPDate(etag.Time),
 				ETag:         etag.String(),
 			},
 		}, nil
@@ -106,7 +106,7 @@ func (c Categories) CategoryGet(ctx context.Context, request openapi.CategoryGet
 			Body: serialiseCategory(cat),
 			Headers: openapi.CategoryGetOKResponseHeaders{
 				CacheControl: categoryGetCacheControl,
-				LastModified: httpDate(etag.Time),
+				LastModified: cachecontrol.HTTPDate(etag.Time),
 				ETag:         etag.String(),
 			},
 		},
