--- app/transports/http/bindings/httpdate.go
@@ -1,10 +1,0 @@
-package bindings
-
-import (
-	"net/http"
-	"time"
-)
-
-func httpDate(t time.Time) string {
-	return t.UTC().Format(http.TimeFormat)
-}
